moved function to server
